+++ pyproject.toml
@@ -32,7 +32,6 @@ dependencies = [
   "moviepy>=1.0.3",
   "tensordict==0.2.*",
   "tensorboard>=2.10",
-  "mlflow==2.8.0",
   "python-dotenv>=1.0.0",
   "lightning==2.1.*",
   "lightning-utilities<=0.9",
@@ -84,6 +83,7 @@ minedojo = ["minedojo==0.1", "importlib_resources==5.12.0"]
 minerl = ["setuptools==66.0.0", "minerl==0.4.4"]
 diambra = ["diambra==0.0.16", "diambra-arena==2.2.2"]
 crafter = ["crafter==1.8.1"]
+mlflow = ["mlflow==2.8.0"]
 
 [tool.ruff]
 line-length = 120
